Moved testutils inside internal package
