+++ internal/testutils/test_container_utils.go
@@ -1,4 +1,4 @@
-package utils_test
+package testutils
 
 import (
 	"context"
